19165: Adds `exist_ok=True` to trace directory creation call (#53)
Adding `exist_ok=True` on the `mkdir` call that creates the directory
for the tracefile.

We believe this may be the cause of some test failures where a race
condition exists where multiple processes find the directory does not
exist, then one creates the directory before the rest. This means that
when the following processes attempt to create the directory, it already
existed. Adding this flag should prevent these exceptions.