21160: Adds retries to PyPi publish, reorders release steps (#111)
The step in the release workflow that publishes to PyPi often fails on
the first attempt and requires a retry. However, the GitHub release
happens _before_ the PyPi release currently, so someone has to go back
and delete the GitHub release and tag and then retry. This reorders the
steps to do the GitHub release _after_ a successful PyPi release, and
also adds a retry wrapper to the PyPi release (3 attempts).