24630: load/store bytes, entities, MINOR (#201)
Add new Amalgam.load_entity_bytes() and store_entity_bytes() methods
that accept a Python bytes object instead of a file object.  This
allows loading and storing an entity without landing it on disk.

Add an entity_path parameter to the four "load" and "store" methods.
This is a list of strings that allows loading or storing an entity
within a nested tree of Amalgam entities.